backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
aiofiles==24.1.0
anthropic==0.40.0
openai==1.54.0
fal-client==0.5.6
stability-sdk==0.8.4
google-generativeai==0.8.3
PyMuPDF==1.24.14
playwright==1.49.0
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
